versiones definitivas

+++ reto-07/Joinzar/src/utils.py
@@ -88,20 +88,14 @@ def list_images(directory):
             pass
 
 
-def copy_files(directory_in, directory_out):
+def copy_files(directory_in, directory_out, filter_fil="filtro"):
     """ Función copiar archivos desde directory_in a directory_out
-
-    Asumo que la función es copiar todos los archivos que existan en el
-    directorio_in al out y que el escenario es que existan sólo archivos,
-    ya que si existieran directorios, el escenario cambia.
-
+    Copiará los archivos que respondan al filtro dado.
     Se comprueba que los directorios existan. Si in no existe, imprime
     aviso y cesa. Si out no existe, se crea.
-
     Para cada archivo, debe comprobar que no exista un archivo con el
     mismo nombre en out y si no existe, copiarlo.
-    Meto un mensaje de confirmación (verbose)
-
+    Meto un mensaje de confirmación (verbose).
     """
     if os.path.exists(directory_in):
         if not os.path.exists(directory_out):
@@ -109,31 +103,33 @@ def copy_files(directory_in, directory_out):
         ficheros_in = os.listdir(directory_in)
         ficheros_out = os.listdir(directory_out)
         for afile in ficheros_in:
-            if afile not in ficheros_out:
-                shutil.copy(directory_in + '/' + afile, directory_out)
-            else:
-                # print(f"{afile} ya existe, no se copiará")
-                pass
+            """ Trabajo en la hipótesis de que los filtros sólo pueden ser
+            o bien *.extension o *.*, esto es, cualquiera"""
+            if ((filter_fil.partition('.')[-2] +
+                    filter_fil.partition('.')[-1]) in afile
+                    or filter_fil == '*.*'):
+                if afile not in ficheros_out:
+                    shutil.copy(directory_in + '/' + afile, directory_out)
+                else:
+                    # print(f"{afile} ya existe, no se copiará")
+                    pass
     else:
         # print(f"{directory_in} no existe. Nada para copiar")
         pass
 
 
-def move_files(directory_in, directory_out):
+def move_files(directory_in, directory_out, filter_fil="filtro"):
     """ Función mover archivos desde directory_in a directory_out
-
     Asumo que la función es mover todos los archivos que existan en el
-    directorio_in al out y que el escenario es que existan sólo archivos,
-    ya que si existieran directorios, el escenario cambia.
-
+    directorio_in al out que respondan al filtro dado y que el escenario
+    es que existan sólo archivos, ya que si existieran directorios,
+    el escenario cambia.
     Se comprueba que los directorios existan. Si in no existe, levantaría
     error y cesa. Si out no existe, se crea.
-
     Según instrucciones, se deben mover los archivos de in a out,
     pero si existen en destino, se deben borrar antes. ¿Cuáles?
     Entiendo que para cada archivo.
     Meto un mensaje de confirmación (verbose)
-
     """
     if os.path.exists(directory_in):
         if not os.path.exists(directory_out):
@@ -141,11 +137,16 @@ def move_files(directory_in, directory_out):
         ficheros_in = os.listdir(directory_in)
         ficheros_out = os.listdir(directory_out)
         for afile in ficheros_in:
-            if afile not in ficheros_out:
-                shutil.move(directory_in + '/' + afile, directory_out)
-            else:
-                os.remove(directory_out + '/' + afile)
-                shutil.move(directory_in + '/' + afile, directory_out)
+            """ Trabajo en la hipótesis de que los filtros sólo pueden ser
+            o bien *.extension o *.*, esto es, cualquiera"""
+            if ((filter_fil.partition('.')[-2] +
+                    filter_fil.partition('.')[-1]) in afile
+                    or filter_fil == '*.*'):
+                if afile not in ficheros_out:
+                    shutil.move(directory_in + '/' + afile, directory_out)
+                else:
+                    os.remove(directory_out + '/' + afile)
+                    shutil.move(directory_in + '/' + afile, directory_out)
     else:
         # print(f"{directory_in} no existe. Nada para copiar")
         pass
